+++ bin/kanban.js
@@ -1,0 +1,12 @@
+#!/usr/bin/env node
+
+const path = require('path');
+
+const { main } = require(path.join(__dirname, '..', 'get-shit-done', 'tools', 'kanban.js'));
+
+main(process.argv).catch((err) => {
+  // eslint-disable-next-line no-console
+  console.error(err);
+  process.exitCode = 1;
+});
+